requirements.txt
alembic
apscheduler
fastapi
httpx
langdetect
ollama
pydantic-settings
pydantic_core==2.41.5
pymupdf
sqlalchemy
uvicorn
